Secured views so only logged in users and superusers can view relevant pages

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 
 from .models import Product, Colour, Category, Style, Material, Backing, Manufacturer
@@ -95,8 +96,14 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add Prooduct to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to view this page!')
+        return redirect(reverse('home'))
+
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -129,8 +136,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to view this page!')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -153,7 +165,12 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to view this page!')
+        return redirect(reverse('home'))
+
     """Delete Product"""
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
